+++ assets/prototype/application/ui/components/display/entityList/filterBar/Collapsible.js
