+++ assets/prototype/application/ui/components/display/entityList/filterBar/EntityListFilterBar.js
