+++ assets/prototype/domain/eventEditor/datetimes/dateList/DateList.js
@@ -5,7 +5,7 @@ import EmptyState from '../../../../application/ui/components/display/EmptyState
 import useDatetimes from '../../data/queries/datetimes/useDatetimes';
 import useStatus from '../../../../application/services/apollo/status/useStatus';
 
-import DatesListFilterBar from './filter/DatesListFilterBar';
+import DatesListFilterBar from './filterBar/DatesListFilterBar';
 import AddNewDateButton from './AddNewDateButton';
 import DateCard from '../dateCard/DateCard';
 
@@ -62,7 +62,7 @@ const DateList = () => {
 	return (
 		<div>
 			{header}
-			<DatesListFilterBar datetimes={datetimes} />
+			<DatesListFilterBar />
 			{datesList}
 		</div>
 	);
