+++ assets/prototype/domain/eventEditor/datetimes/dateList/filterBar/controls/ShowDatesControl.js
